update waterParam function, fix error

+++ kcs-backend/src/main/java/group7/se1876/kcs_backend/service/FishService.java
@@ -7,15 +7,15 @@ import group7.se1876.kcs_backend.dto.response.FishResponse;
 import group7.se1876.kcs_backend.dto.response.KoiFishDevelopmentResponse;
 import group7.se1876.kcs_backend.entity.Fish;
 import group7.se1876.kcs_backend.entity.FishDevelopmentHistory;
+import group7.se1876.kcs_backend.entity.Pond;
 import group7.se1876.kcs_backend.entity.User;
 import group7.se1876.kcs_backend.exception.AppException;
 import group7.se1876.kcs_backend.exception.ErrorCode;
 import group7.se1876.kcs_backend.mapper.FishMapper;
-import group7.se1876.kcs_backend.repository.FishHistories;
+import group7.se1876.kcs_backend.repository.FishHistoriesRepository;
 import group7.se1876.kcs_backend.repository.FishRepository;
 import group7.se1876.kcs_backend.repository.UserRepository;
 import lombok.AllArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
@@ -32,7 +32,7 @@ public class FishService {
     private FishRepository fishRepository;
     private FishMapper fishMapper;
     private UserRepository userRepository;
-    private FishHistories fishHistories;
+    private FishHistoriesRepository fishHistories;
 
 
     //Add fish
@@ -68,6 +68,10 @@ public class FishService {
 
         Fish fish = fishRepository.findById(fishId)
                 .orElseThrow(()->new AppException(ErrorCode.DATA_NOT_EXISTED));
+        Long userId = Long.valueOf(SecurityContextHolder.getContext().getAuthentication().getName());
+
+        if (userId!=fish.getOwner().getUserId())
+            throw new AppException(ErrorCode.INVALID_USERID);
 
         return fishMapper.mapToFishResponse(fish);
     }
@@ -77,6 +81,10 @@ public class FishService {
 
         Fish fish = fishRepository.findById(fishId)
                 .orElseThrow(()->new AppException(ErrorCode.DATA_NOT_EXISTED));
+
+        Long userId = Long.valueOf(SecurityContextHolder.getContext().getAuthentication().getName());
+        if (userId!=fish.getOwner().getUserId())
+            throw new AppException(ErrorCode.INVALID_DATA_WITH_USERID);
 
         fish.setFishName(request.getFishName());
         fish.setFishImg(request.getFishImg());
@@ -103,15 +111,24 @@ public class FishService {
                 .orElseThrow(()-> new AppException(ErrorCode.DATA_NOT_EXISTED));
 
         User user = fish.getOwner();
+        Long userId = Long.valueOf(SecurityContextHolder.getContext().getAuthentication().getName());
 
-        try {
-            if (user!=null){
-                user.getFishOwned().remove(fish);
-                userRepository.save(user);
-            }
-        }catch (Exception e){
-            throw new AppException(ErrorCode.DELETE_FAIL);
+
+        if (userId!=fish.getOwner().getUserId())
+            throw new AppException(ErrorCode.INVALID_DATA_WITH_USERID);
+
+        for (Pond pond : fish.getPonds()) {
+            pond.getFish().remove(fish);  // Remove fish from the pond
         }
+
+        if (user != null) {
+            // Remove the fish from the user's list
+            user.getFishOwned().remove(fish);
+            // Save the updated user object
+            userRepository.save(user);
+        }
+
+
         fishRepository.delete(fish);
 
     }
@@ -125,7 +142,7 @@ public class FishService {
         Long userId = Long.valueOf(SecurityContextHolder.getContext().getAuthentication().getName());
 
         if (!fish.getOwner().getUserId().equals(userId)){
-            throw new AppException(ErrorCode.INVALID_USERID);
+            throw new AppException(ErrorCode.INVALID_DATA_WITH_USERID);
         }
 
         FishDevelopmentHistory fishDevelopmentHistory = fishMapper.mapToFishHistory(request) ;
@@ -136,5 +153,24 @@ public class FishService {
 
         return fishMapper.mapToKoiFishResponse(fishDevelopmentHistory);
 
+    }
+
+    //Get all fish history
+    public List<KoiFishDevelopmentResponse> getFishHistories(Long fishId){
+
+        Fish fish = fishRepository.findById(fishId)
+                .orElseThrow(()-> new AppException(ErrorCode.DATA_NOT_EXISTED));
+
+        Long userId = Long.valueOf(SecurityContextHolder.getContext().getAuthentication().getName());
+
+        if (!fish.getOwner().getUserId().equals(userId)){
+            throw new AppException(ErrorCode.INVALID_DATA_WITH_USERID);
+        }
+
+        List<FishDevelopmentHistory> fishDevelopmentHistories = fish.getFishDevelopmentHistories();
+
+        return fishDevelopmentHistories.stream()
+                .map((fishDevelopmentHistory)-> fishMapper.mapToKoiFishResponse(fishDevelopmentHistory))
+                .collect(Collectors.toList());
     }
 }
